backend/requirements.txt
# core AWS SDK
boto3==1.38.41
botocore==1.38.41
s3transfer==0.13.0
jmespath==1.0.1
six==1.17.0
urllib3==2.5.0
python-dateutil==2.9.0.post0
numpy==1.24.4
python-dotenv

# web API
fastapi
mangum

# data science
pandas


# database connector
snowflake-connector-python
# Subir archivos mediante fastapi
python-multipart